backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
psycopg2-binary>=2.9.6
celery>=5.3.0
redis>=4.5.5
requests>=2.30.0
python-dotenv>=1.0.0
flower>=2.0.0
sqlalchemy>=2.0.0
email-validator>=2.0.0
fastapi-limiter>=0.1.5
pydantic-settings>=2.0.0
psutil>=5.9.0
aioredis>=2.0.0
boto3>=1.24.0